requirements.txt
openai==0.28.0
Pillow==10.1.0
streamlit==1.29.0
